2.2 fixing up rare case naming conventions for plugins

+++ src/main/java/org/grails/maven/plugin/GrailsPackageBinaryPluginMojo.java
@@ -64,16 +64,24 @@ public class GrailsPackageBinaryPluginMojo extends AbstractGrailsMojo {
     // Now move the JAR from the project directory to the build
     // output directory.
     final String mavenFileName = project.getArtifactId() + "-" + project.getVersion() + ".jar";
-    String jarFileName = PLUGIN_PREFIX + "plugin-" + mavenFileName;
+    String jarFileName = PLUGIN_PREFIX + "plugin-";
+    if (!mavenFileName.startsWith(PLUGIN_PREFIX))
+      jarFileName += mavenFileName;
+    else {
+      jarFileName += mavenFileName.substring(PLUGIN_PREFIX.length());
+    }
 //
     File fileGeneratedByGrails = new File(project.getBuild().getDirectory(), jarFileName);
 //
     File existingJarFile = new File(project.getBuild().getDirectory(), mavenFileName);
-    existingJarFile.delete();
-    if (!fileGeneratedByGrails.renameTo(existingJarFile)) {
-      throw new MojoExecutionException("Unable to copy the plugin binary JAR to the target directory (" + fileGeneratedByGrails.getAbsolutePath() + " to " + existingJarFile.getAbsolutePath() + ")");
-    } else {
-      getLog().info("Moved plugin binary JAR to '" + existingJarFile + "'.");
+
+    if (!fileGeneratedByGrails.equals(existingJarFile)) {
+      existingJarFile.delete();
+      if (!fileGeneratedByGrails.renameTo(existingJarFile)) {
+        throw new MojoExecutionException("Unable to copy the plugin binary JAR to the target directory (" + fileGeneratedByGrails.getAbsolutePath() + " to " + existingJarFile.getAbsolutePath() + ")");
+      } else {
+        getLog().info("Moved plugin binary JAR to '" + existingJarFile + "'.");
+      }
     }
 
     artifact.setFile(existingJarFile);
